v2.0

+++ Assets/Scripts/Utilities/PersistentData.cs
@@ -7,11 +7,11 @@ public static class PersistentData {
     //Every information that will be stored should have it's set/get
 
     /*  Highscore set/get   */
-    public static void SetHighscore(float newHighscore) {
-        PlayerPrefs.SetFloat("Highscore", newHighscore);
+    public static void SetHighscore(int newHighscore) {
+        PlayerPrefs.SetInt("Highscore", newHighscore);
     }
-    public static float GetHighscore() {
-        return PlayerPrefs.GetFloat("Highscore");
+    public static int GetHighscore() {
+        return PlayerPrefs.GetInt("Highscore");
     }
     public static void SetUserRankCoeficient(float newCoeficient) {
         PlayerPrefs.SetFloat("userRankCoeficient", newCoeficient);
